Move indexing functions to supabase project
