+++ .vscode/settings.json
@@ -1,5 +1,6 @@
 {
   "typescript.tsdk": "node_modules/typescript/lib",
   "typescript.enablePromptUseWorkspaceTsdk": true,
-  "deno.enablePaths": ["./supabase"]
+  "deno.enablePaths": ["./supabase"],
+  "deno.importMap": "./supabase/functions/import_map.json"
 }
