+++ supabase/functions/import_map.json
@@ -1,0 +1,6 @@
+{
+  "imports": {
+    "ethers": "https://esm.sh/ethers@6",
+    "@supabase/supabase-js": "https://esm.sh/@supabase/supabase-js@2"
+  }
+}
