+++ supabase/functions/utils/indexing/pods.ts
@@ -1,12 +1,12 @@
 import { ethers } from "ethers";
 import { SupabaseClient } from "@supabase/supabase-js";
-import { rangeChunkMap } from "./utils/chunk";
-import { EIGEN_POD_MANAGER_ADDRESS, INDEXING_BLOCK_CHUNK_SIZE } from "./utils/constants";
-import { getIndexingEndBlock, getIndexingStartBlock, releaseBlockLock, setLastIndexedBlock } from "./utils/updates";
-import { EigenPodManager__factory } from "../../typechain";
+import { rangeChunkMap } from "./utils/chunk.ts";
+import { EIGEN_POD_MANAGER_ADDRESS, INDEXING_BLOCK_CHUNK_SIZE } from "./utils/constants.ts";
+import { getIndexingEndBlock, getIndexingStartBlock, setLastIndexedBlock } from "./utils/updates.ts";
+import { EigenPodManager__factory } from "../../typechain/index.ts";
 
 // serialization polyfill
-import "./utils/bigint";
+import "./utils/bigint.ts";
 
 interface Pod {
   block: number;
