+++ supabase/functions/utils/indexing/utils/beaconProvider.ts
