+++ supabase/functions/utils/indexing/utils/constants.ts
