+++ supabase/functions/utils/indexing/utils/trace.ts
